refactor: move `SnackBarData` and `LocalSnackBarController` to separate files
