+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/components/snackBar/IdentitySnackBarController.kt
@@ -1,37 +1,14 @@
 package net.thechance.mena.identity.presentation.components.snackBar
 
-import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import androidx.compose.runtime.staticCompositionLocalOf
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.receiveAsFlow
 import mena.identity_presentation.generated.resources.Res
 import mena.identity_presentation.generated.resources.error
-import mena.identity_presentation.generated.resources.ic_close_circle
-import mena.identity_presentation.generated.resources.ic_success
 import mena.identity_presentation.generated.resources.success
-import org.jetbrains.compose.resources.DrawableResource
 import org.jetbrains.compose.resources.StringResource
-
-val LocalSnackBarController = staticCompositionLocalOf<IdentitySnackBarController> {
-    error("No SnackBarController provided")
-}
-
-@Immutable
-data class SnackBarData(
-    val message: StringResource,
-    val title: StringResource,
-    val duration: Long = 3_000L,
-    val type: SnackBarType = SnackBarType.ERROR
-) {
-
-    enum class SnackBarType(val icon: DrawableResource) {
-        ERROR(Res.drawable.ic_close_circle),
-        SUCCESS(Res.drawable.ic_success),
-    }
-}
 
 class IdentitySnackBarController() {
