+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/components/snackBar/LocalSnackBarController.kt
@@ -1,0 +1,7 @@
+package net.thechance.mena.identity.presentation.components.snackBar
+
+import androidx.compose.runtime.staticCompositionLocalOf
+
+val LocalSnackBarController = staticCompositionLocalOf<IdentitySnackBarController> {
+    error("No SnackBarController provided")
+}
